DsToMT5/dsbot/requirements.txt
discord.py-self
python-telegram-bot
requests